Fix: forgotten uint256 indexed chainId fields

+++ packages/contracts-communication/contracts/events/SynapseGasOracleV1Events.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.0;
 
 abstract contract SynapseGasOracleV1Events {
-    event CalldataPriceSet(uint256 indexed chainId, uint256 calldataPrice);
-    event GasPriceSet(uint256 indexed chainId, uint256 gasPrice);
-    event NativePriceSet(uint256 indexed chainId, uint256 nativePrice);
+    event CalldataPriceSet(uint64 indexed chainId, uint256 calldataPrice);
+    event GasPriceSet(uint64 indexed chainId, uint256 gasPrice);
+    event NativePriceSet(uint64 indexed chainId, uint256 nativePrice);
 }
